Move validate_synthetics_transaction_trace into validators directory
